Move unit tests into integration tests folder
